refactor: remove `_interchange.series.py`
If two modules are dependent on one another,
they are one module
