+++ src/narwhals/_interchange/__init__.py
@@ -1,6 +1,10 @@
 from __future__ import annotations
 
-from narwhals._interchange.dataframe import InterchangeFrame, should_interchange
+from narwhals._interchange.dataframe import (
+    InterchangeFrame,
+    InterchangeSeries,
+    should_interchange,
+)
 from narwhals._interchange.lazyframe import LazyFrame
 
-__all__ = "InterchangeFrame", "LazyFrame", "should_interchange"
+__all__ = "InterchangeFrame", "InterchangeSeries", "LazyFrame", "should_interchange"
